+++ helm/bin/helm-save.sh
@@ -11,15 +11,25 @@ then
     exit 1
 fi
 
+umask 022
+
 FILE="../${1#*/}_${2}_app-${3}.tar"
 shift 3
 LIST="${FILE%.tar}.list"
 CREATED=$(docker inspect "$@" | sed -e 's/"/ /g' | sort -r | awk '/Created/ {print $3; exit}')
 
-echo "docker save $@ -o $FILE"
-docker save "$@" -o "$FILE.part"
-tar xf "$FILE.part" repositories -O | sed -e "s/,/\\n/g;s/\"/:/g" | cut -f2,5 -d: > "$LIST.part"
-touch -d $CREATED "$FILE.part" "$LIST.part"
-chmod 644 "$FILE.part"
-mv "$FILE.part" "$FILE"
-mv "$LIST.part" "$LIST"
+if [ -f "$FILE.7z" ]
+then
+    echo "Skipping existing file: $(realpath $FILE.7z)"
+else
+    echo "docker save $@ -o $FILE"
+    docker save "$@" -o "$FILE.part"
+    tar xf "$FILE.part" repositories -O | sed -e "s/,/\\n/g;s/\"/:/g" | cut -f2,5 -d: > "$LIST.part"
+    touch -d $CREATED "$FILE.part" "$LIST.part"
+    mv "$FILE.part" "$FILE"
+    mv "$LIST.part" "$LIST"
+    7z a -m0=lzma2 -mmt=2 -mx=9 -myx=9 -md=128m -mfb=256 -ms=on -snh -snl -stl -y "$FILE.7z.part" "$FILE"
+    mv "$FILE.7z.part" "$FILE.7z"
+    rm "$FILE"
+    echo "Created archive file: $(realpath $FILE.7z)"
+fi
